Make use of INavigationService for Navigation from VM where applicable
