+++ WingetGUIInstaller/Enums/NavigationItemKey.cs
@@ -10,5 +10,6 @@
         Search,
         Settings,
         Upgrades,
+        PackageDetails,
     }
 }
