+++ WingetGUIInstaller/Pages/ListPage.xaml.cs
@@ -19,17 +19,5 @@ namespace WingetGUIInstaller.Pages
             DataContext = ViewModel = Ioc.Default.GetRequiredService<ListPageViewModel>();
             InitializeComponent();
         }
-
-        public ICommand GoToDetailsCommand =>
-            new RelayCommand<PackageDetailsViewModel>(ViewPackageDetails);
-
-        private void ViewPackageDetails(PackageDetailsViewModel obj)
-        {
-            Frame.Navigate(typeof(PackageDetailsPage), new PackageDetailsNavigationArgs
-            {
-                PackageDetails = obj,
-                AvailableOperation = AvailableOperation.Uninstall
-            });
-        }
     }
 }
