+++ WingetGUIInstaller/Pages/PackageDetailsPage.xaml.cs
@@ -1,20 +1,26 @@
 ﻿using CommunityToolkit.Mvvm.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
+using WingetGUIInstaller.Enums;
 using WingetGUIInstaller.Models;
+using WingetGUIInstaller.Services;
+using WingetGUIInstaller.Utils;
 using WingetGUIInstaller.ViewModels;
 
 namespace WingetGUIInstaller.Pages
 {
+    [PageKey(NavigationItemKey.PackageDetails)]
     public sealed partial class PackageDetailsPage : Page
     {
+        private readonly INavigationService<NavigationItemKey> _navigationService;
+
         internal PackageDetailsPageViewModel ViewModel { get; }
 
         public PackageDetailsPage()
         {
             InitializeComponent();
+            _navigationService = Ioc.Default.GetRequiredService<INavigationService<NavigationItemKey>>();
             DataContext = ViewModel = Ioc.Default.GetRequiredService<PackageDetailsPageViewModel>();
             NavigationCacheMode = NavigationCacheMode.Disabled;
         }
@@ -31,7 +37,7 @@ namespace WingetGUIInstaller.Pages
 
         private void GoBack_Click(object sender, RoutedEventArgs e)
         {
-            Frame.GoBack();
+            _navigationService.GoBack();
         }
     }
 }
