+++ WingetGUIInstaller/Pages/RecommendationsPage.xaml.cs
@@ -2,18 +2,22 @@
 using Microsoft.UI.Xaml.Controls;
 using WingetGUIInstaller.Enums;
 using WingetGUIInstaller.Models;
+using WingetGUIInstaller.Services;
 using WingetGUIInstaller.Utils;
 using WingetGUIInstaller.ViewModels;
 
 namespace WingetGUIInstaller.Pages
 {
-    [PageKey(Enums.NavigationItemKey.Recommendations)]
+    [PageKey(NavigationItemKey.Recommendations)]
     public sealed partial class RecommendationsPage : Page
     {
+        private readonly INavigationService<NavigationItemKey> _navigationService;
+
         public RecommendationsPageViewModel ViewModel { get; }
 
         public RecommendationsPage()
         {
+            _navigationService = Ioc.Default.GetRequiredService<INavigationService<NavigationItemKey>>();
             DataContext = ViewModel = Ioc.Default.GetRequiredService<RecommendationsPageViewModel>();
             InitializeComponent();
         }
@@ -22,7 +26,7 @@ namespace WingetGUIInstaller.Pages
         {
             if (e.ClickedItem is RecommendedItemViewModel recommendedItem)
             {
-                Frame.Navigate(typeof(PackageDetailsPage), new PackageDetailsNavigationArgs
+                _navigationService.Navigate(NavigationItemKey.PackageDetails, new PackageDetailsNavigationArgs
                 {
                     AvailableOperation = recommendedItem.IsInstalled ? AvailableOperation.Uninstall : AvailableOperation.Install,
                     PackageId = recommendedItem.Id,
