+++ WingetGUIInstaller/Pages/SearchPage.xaml.cs
@@ -19,17 +19,5 @@ namespace WingetGUIInstaller.Pages
             DataContext = ViewModel = Ioc.Default.GetRequiredService<SearchPageViewModel>();
             InitializeComponent();
         }
-
-        public ICommand GoToDetailsCommand =>
-            new RelayCommand<PackageDetailsViewModel>(ViewPackageDetails);
-
-        private void ViewPackageDetails(PackageDetailsViewModel obj)
-        {
-            Frame.Navigate(typeof(PackageDetailsPage), new PackageDetailsNavigationArgs
-            {
-                PackageDetails = obj,
-                AvailableOperation = AvailableOperation.Install
-            });
-        }
     }
 }
